Re #112 Refactor systemtest helper functions into seperate script
