+++ system-tests/docker-compose-fake-epics.yml
@@ -23,15 +23,6 @@ services:
     image: zookeeper:3.4
     restart: unless-stopped
 
-  ioc:
-    image: dmscid/lewis:latest
-    network_mode: "host"
-    depends_on:
-      - zookeeper
-      - kafka
-    restart: unless-stopped
-    command: "chopper -p \"epics: {prefix: 'SIM:'}\""
-
   forwarder:
     build: ../
     depends_on:
