+++ system-tests/test_generate_fake_epics_data.py
@@ -1,2 +1,3 @@
-def test_something(docker_compose_fake_epics):
+def test_forwarder_sends_fake_pv_updates(docker_compose_fake_epics):
+    # We should see PV updates in Kafka despite there being no IOC running
     pass
